Refactor character limit validation utilities: consolidate and enhance functionality
